Assistant checkpoint: Add Investment Metrics tables to PDF report
Assistant generated file changes:
- client/src/features/property-analyzer-pdf/services/PDFService.ts: Add Investment Metrics tables to PDF

---

User prompt:

we need to add a new section into the PDF called Investment metrics. It should be all the same metrics used in the investment metrics component on the property analyser page. They can be added using the same methodology as the cashflow metrics. But remember we're not recalcualting anything but just taking the data that has been returned by the property analyser. We need two of these tables, one for long term and one for short term.