Assistant checkpoint: Improve limit reached UX with modal dialog
Assistant generated file changes:
- client/src/pages/PropertyAnalyzerPage.tsx: Improve limit reached UX with modal

---

User prompt:

This is a shit way of telling the user they are out of free analyses. We need to load the page and then check if user has hit their limit. If yes, grey out the whole page and display a popup explaining to the user that they need to upgrade. Remove the notfication link to the analyse button and instead have it on the property analyser page as a whole.