Assistant checkpoint: Added analyses counter to property analyzer page
Assistant generated file changes:
- client/src/pages/PropertyAnalyzerPage.tsx: Add analyses counter display

---

User prompt:

In the top right of the property analyser page, please add a static piece of text. It should says "Analyses used: 0". This is going to be used to display a counter of how many times the user has clicked the analyse button in the property analyser form. Make sure to use good ui and design, and use the proply blue where necessary.