Assistant checkpoint: Add upgrade buttons with modal for pro templates
Assistant generated file changes:
- client/src/features/property-analyzer-pdf/components/PDFGenerator.tsx: Add upgrade buttons for pro templates, Add upgrade modal state, Add UpgradeModal component

---

User prompt:

See the dropdown in the screenshot?  could we add a little button next to each of the other 2 report templates called upgrade. This is in the dropdown and should only be visible to free users. And it should then bring up the upgrade modal component if clicked. It can even be small underlined blue text saying upgrade.