Assistant checkpoint: Add pro badge and logo upload functionality to branding dialog
Assistant generated file changes:
- client/src/components/PropertyPreviewModal.tsx: Add pro badge and logo upload functionality, Add required imports

---

User prompt:

Ok now, next to the yes button (or inside it) we need to add the pro badge as this will be a pro feature. If the user clicks no, it should just close the dialog and create the pdf. If they user clicks yes, it should pull the image that has already been uploaded in the profile section of the settings page. It is possible the user hasnt uploaded this logo/image yet. In that case they should be able to upload a logo in this current dialog and that should then save to their profile section in the settings page. Does that all make sense?